+++ test/bin/docker-run.sh
@@ -1,8 +1,5 @@
 #!/bin/bash
 
-# Start MongoDB
-/sbin/start-mongodb.sh
-
 export CI=true
 export CIRCLECI=true
 export PATH="$HOME/.rbenv/bin:$PATH"